Aggiunto aceesso automatico alla posizione del personaggio, abbozzata funzione generica di accesso a posizione di un'entità.

+++ Nella oscura e pronda caverna/Nella oscura e pronda caverna/Attributi.cpp
@@ -40,7 +40,7 @@ Attributi::Attributi(int forza, int destrezza, int tempra, int intelligenza, int
 	hp = maxHp;
 	potereMagico = maxPotereMagico;
 	stamina = maxStamina;
-	forzaDiCarico = (forza * 10 + 1)*(forza / 8);
+	forzaDiCarico = (forza * 10 + 1)*(double)(forza / 8.);
 	if (destrezza > 7 && livello > 10) { //Segnala se l'entità è percepibile aldilà della vista
 		silenzioso = true;
 	}
